+++ saleor/asgi/gunicorn_worker.py
@@ -2,6 +2,10 @@ from uvicorn.workers import UvicornWorker as BaseUvicornWorker
 
 
 class UvicornWorker(BaseUvicornWorker):
-    # Override default config args, that cannot be passed as gunicorn parameters,
-    # to disable lifespan, not supported by django
-    CONFIG_KWARGS = {"loop": "uvloop", "http": "httptools", "lifespan": "off"}
+    # Override default config args, that cannot be passed as gunicorn parameters.
+    CONFIG_KWARGS = {
+        "loop": "uvloop",
+        "http": "httptools",
+        "lifespan": "off",
+        "ws": "none",
+    }